[REFACTOR] ParkingOption enum 삭제
이유: 옵션아이템으로 대체
