--- src/main/java/org/hanihome/hanihomebe/property/domain/enums/ParkingOption.java
@@ -1,7 +1,0 @@
-package org.hanihome.hanihomebe.property.domain.enums;
-
-public enum ParkingOption {
-    NONE,
-    RESERVED_SPACE,
-    STREET_PARKING
-}
